requirements.txt
aiohttp==3.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coincurve==8.0.2
fastnumbers==2.1.0
Flask==1.0.2
idna==2.7
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
multidict==4.3.1
py-ecc==1.4.3
pycparser==2.18
pycryptodome==3.6.4
python-idex==0.3.3
requests==2.19.1
rlp==0.4.7
urllib3==1.23
websockets==6.0
Werkzeug==0.14.1
yarl==1.2.6
